requirements.txt
requests
python-telegram-bot
tabulate
bunch
boto3
fuzzywuzzy
python-Levenshtein
cachetools
pandas
matplotlib
seaborn
arrow
pytz
jmespath